backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0

# PDF Processing
PyPDF2==3.0.1

# AI/ML for semantic matching
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch==2.1.1
